# docs/contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ArquitecturaBaseMultitenant.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/me/external/google": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleChallengeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleChallengeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleChallengeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/login-methods": {
      "get": {
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AccountLoginMethodsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountLoginMethodsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountLoginMethodsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLoginEmailHttpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLoginEmailHttpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddLoginEmailHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/login-methods/{methodId}/code": {
      "post": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "methodId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMethodCodeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/login-methods/{methodId}/verify": {
      "post": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "methodId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyLoginMethodHttpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyLoginMethodHttpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyLoginMethodHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/login-methods/{methodId}/primary": {
      "put": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "methodId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/login-methods/{methodId}": {
      "delete": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "name": "methodId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeLoginMethodHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountLoginMethodResponse": {
        "required": [
          "id",
          "type",
          "isPrimary",
          "isVerified",
          "canRemove",
          "canMakePrimary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/LoginMethodType"
          },
          "value": {
            "type": [
              "null",
              "string"
            ]
          },
          "isPrimary": {
            "type": "boolean"
          },
          "isVerified": {
            "type": "boolean"
          },
          "managedByTenantId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "managedByOrganizationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "canRemove": {
            "type": "boolean"
          },
          "canMakePrimary": {
            "type": "boolean"
          },
          "backupDestination": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Describe disponibilidad y acciones de un método sin incluir su contacto completo en la representación de diagnóstico."
      },
      "AccountLoginMethodsResponse": {
        "required": [
          "methods",
          "canLinkGoogle",
          "needsPersonalLoginMethod",
          "accountDeletionGraceDays"
        ],
        "type": "object",
        "properties": {
          "methods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountLoginMethodResponse"
            }
          },
          "canLinkGoogle": {
            "type": "boolean"
          },
          "needsPersonalLoginMethod": {
            "type": "boolean"
          },
          "accountDeletionGraceDays": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Entrega métodos de ingreso y capacidades de gestión de la cuenta sin exponer los objetos de persistencia."
      },
      "AddLoginEmailHttpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "reauthTicket": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Recibe el correo que se quiere agregar a la cuenta y lo entrega al flujo de validación. ToString oculta\r\nlos datos sensibles para que no se impriman al registrar la solicitud."
      },
...
      "ChangeLoginMethodHttpRequest": {
        "type": "object",
        "properties": {
          "reauthTicket": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Recibe la prueba de titularidad para quitar o elegir como principal un método de ingreso. El método\r\nobjetivo llega en la ruta; el ticket se conserva sin normalizar y se oculta en ToString."
      },
...
      "GoogleChallengeResponse": {
        "required": [
          "redirectUrl"
        ],
        "type": "object",
        "properties": {
          "redirectUrl": {
            "type": "string"
          }
        },
        "description": "Entrega la URL del desafío Google al controller sin exponerla en la representación textual del objeto."
      },
...
      "LoginMethodCodeResponse": {
        "required": [
          "methodId",
          "resendAfterSeconds"
        ],
        "type": "object",
        "properties": {
          "methodId": {
            "type": "string",
            "format": "uuid"
          },
          "resendAfterSeconds": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Devuelve el método al que se envió un código y cuánto hay que esperar para reenviarlo. Permite\r\nidentificar el destino y controlar el botón de reenvío."
      },
...
      "LoginMethodType": {
        "enum": [
          "Email",
          "Phone",
          "Google"
        ],
        "type": "string",
        "description": "Identifica si un método de ingreso usa correo, teléfono o Google. El tipo determina cómo se interpreta\r\nsu valor y qué canal puede verificarlo."
      },
...
      "ProblemDetails": {
        "required": [
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "retryAfter": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "VerifyLoginMethodHttpRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Recibe el código usado para verificar un método de ingreso pendiente. Conserva el texto original para\r\nvalidarlo y oculta su valor en ToString."
      },
...
    }
  },
...
}